server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Erdeni Tsyrendashiev <erdenirf@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.41"
mkdocstrings-python = "^1.12.1"
langchain-core = "^0.3.10"
langchain-text-splitters = "^0.3.0"
langchain-openai = "^0.2.2"
langchain-qdrant = "^0.1.4"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
